add support for Oracle Linux 8

+++ init_env_segment.sh
@@ -95,6 +95,13 @@ if [ -f /etc/os-release ]; then
     # Source the /etc/os-release file to get the system information
     source /etc/os-release
 
+    # Checking for Oracle Linux
+    IS_ORACLE_LINUX=0
+    if [[ "$ID" == "ol" || "$NAME" == *"Oracle Linux"* ]]; then
+        IS_ORACLE_LINUX=1
+        echo "This is Oracle Linux"
+    fi
+
     # Extract the first digit of the VERSION_ID
     first_digit=$(echo "$VERSION_ID" | cut -c1)
 
@@ -113,17 +120,20 @@ if [ -f /etc/os-release ]; then
             # sh init_env.sh single
             ;;
         8)
-            # Operation B
+            # Operation in 8
             echo "This is a operating system with version ID starting with 8."
-            rm -rf /etc/yum.repos.d/*
-            curl -o /etc/yum.repos.d/CentOS-Base.repo https://mirrors.huaweicloud.com/repository/conf/CentOS-8-anon.repo
-            yum clean all
-            yum makecache
-            yum install -y libcgroup-tools
+            if [ $IS_ORACLE_LINUX -ne 1 ]; then
+              rm -rf /etc/yum.repos.d/*
+              curl -o /etc/yum.repos.d/CentOS-Base.repo https://mirrors.huaweicloud.com/repository/conf/CentOS-8-anon.repo
+              yum clean all
+              yum makecache
+            else
+              log_time "Skip set yum repo for Oracle Linux."
+            fi
             # You can add specific commands for Operation B here
             ;;
         9)
-            # Operation C
+            # Operation in 9
             echo "This is a operating system with version ID starting with 9. Executing Operation C."
             # You can add specific commands for Operation C here, such as starting the database cluster deployment
             # bash run.sh multi
